stream_handler/requirements.txt
faust-streaming==0.11.3
sortedcontainers==2.4.0
pydantic-settings==2.12.0
pydantic==2.12.5
pyahocorasick==2.3.0
Faker==39.0.0
rocksdict==0.3.29